+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Renamed HTML to Vanilla JS.
 - Updated dependency vue to 3.5.42.
 - Updated dependency vitest to 4.1.11.
 - Updated dependency vite-plugin-solid to 2.11.14.